--- src/Components/Layouts/index.js
@@ -1,8 +1,0 @@
- import {Header} from  "./Header"
- import Footer from "./Footer"
- import LoginMenu from "./LoginMenu"
- import AccountMenu from "./AccountMenu"
- import PracticeList from "./PracticeList"
- export{
-     Header, Footer, LoginMenu, AccountMenu, PracticeList
- }
