+++ src/Components/NewSession/EndDialog.js
