+++ src/Components/NewSession/ImageDialog.js
