+++ src/styles/App.css
@@ -2,27 +2,12 @@
   text-align: center;
 }
 
-.App-logo {
-  /* animation: App-logo-spin infinite 20s linear; */
-  /* height: 40vmin; */
-  font-size: 25px;
-  font-weight: bold;
-  font-family: Georgia;
-  color: white;
-  text-decoration: none;
+ .App-logo {
+  animation: App-logo-spin infinite 20s linear;
+  height: 40vmin;
 }
 
-.Account-container {
-  flex-grow: 1;
-  text-align: right;
-}
-
-.Paper-container {
-  padding: 20px;
-  margin: 15px;
-}
-
-.App-header {
+ .App-header {
   background-color: #282c34;
   min-height: 100vh;
   display: flex;
@@ -33,11 +18,11 @@
   color: white;
 }
 
-.App-link {
+ .App-link {
   color: #61dafb;
 }
 
-@keyframes App-logo-spin {
+ @keyframes App-logo-spin {
   from {
     transform: rotate(0deg);
   }
